Refactored TS2CAST path (but not classname)
